New class SC_CountTarget counts times structure is visited
SC_CountTarget counts how many times a target conformation is visited. e.g. if the target conformation is the native state of the sequence, the P_nat of the sequence can be calculated. This class inserts into the SC hierarchy such that it inherits from SC_Counting and is now the parent of SC_MinE. There are public functions for reporting the count, the fraction of time the target structure is encountered, and the energy of the target conforomation.

SC_MinE has been modified to inherit from SC_CountTarget. Makefile.am has been modified so that latFold and latFoldVec compile with SC_CountTarget.