Add calculation of a survival sum to SC_CountTarget
This is a sum of survival probabilities up to the current state. The survival probability is defined as exp(-k int_0^now non-native(t) dt). It is the probability of not being degraded up to now. This probability is stored as a sum that is added to each time a new state is added.