1.9.1-13, option to calculate integral of survival
This calculation has nothing to do with monte carlo folding. I am modeling the situation where the probability of the protein to be degraded increases with the number of steps in which the conformation is not the specified conformation in -countTarget. In other words, P_survival = exp[- number of steps in non-target conformation / targetDegradationScale]. P_survival is calculated with every step. The simulation reports (sum P_survival(t) * native(t)) / total steps.