15-draft2: fix so fraction target conf is always <= 1
The number of steps since the target is reached was actually undercounted by 1. In rare scenarios in which the target conformation is reached and maintained 100%, the fraction target conf would be > 1. Usually not a problem because both the numerator and denominator are really large, but a specific trajectory occurred which revealed this problem.