SC_CountTarget correction for starting in native
The way stepsToReachTarget was previously set meant that even if the we started in the target structure conformation and the first MC step is to remain in the starting conformation, stepsToReachTarget would end up being 1, when the correct value is 0. The previous way setting stepsToReachTarget is to check if its value is 0. The new way is to see if targetCount == 1, which would be the case for the first MC step that the protein is found to be in the target conformation.