Clear contact list `LatticeProteinUtil::countContacts`
Ok, static variables are bad. `LatticeProteinUtil::countContacts` has a static std::set that holds native contacts. The set could only grow, not shrink, so when `latMapTraj` gets called on a file with multiple trajectories, the calculated native fraction is wrong for all trajectories except the first for intermediate chain lengths. The implemented fix is to `clear()` the set when the length changes, but I won't use static variables in the future.